+++ src/main/java/org/stellar/sdk/TransactionBuilder.java
@@ -10,11 +10,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.function.Function;
 import lombok.NonNull;
-import org.stellar.sdk.TransactionPreconditions.TransactionPreconditionsBuilder;
 import org.stellar.sdk.xdr.SorobanTransactionData;
-import org.stellar.sdk.xdr.TimePoint;
-import org.stellar.sdk.xdr.Uint64;
-import org.stellar.sdk.xdr.XdrUnsignedHyperInteger;
 
 /** Builds a new Transaction object. */
 public class TransactionBuilder {
@@ -24,8 +20,9 @@ public class TransactionBuilder {
   private final List<Operation> operations;
   private Long baseFee;
   private final Network network;
-  private TransactionPreconditions preconditions;
+  @NonNull private TransactionPreconditions preconditions;
   private SorobanTransactionData sorobanData;
+  private BigInteger txTimeout;
 
   /**
    * Construct a new transaction builder.
@@ -153,25 +150,6 @@ public class TransactionBuilder {
   }
 
   /**
-   * Adds a <a href="https://developers.stellar.org/docs/glossary/transactions/"
-   * target="_blank">time-bounds</a> to this transaction.
-   *
-   * @param timeBounds tx can be accepted within this time bound range
-   * @return Builder object so you can chain methods.
-   * @see TimeBounds
-   * @deprecated this method will be removed in upcoming releases, use <code>addPreconditions()
-   *     </code> instead for more control over preconditions.
-   */
-  public TransactionBuilder addTimeBounds(@NonNull TimeBounds timeBounds) {
-    if (preconditions.getTimeBounds() != null) {
-      throw new RuntimeException("TimeBounds already set.");
-    }
-
-    preconditions = preconditions.toBuilder().timeBounds(timeBounds).build();
-    return this;
-  }
-
-  /**
    * Because of the distributed nature of the Stellar network it is possible that the status of your
    * transaction will be determined after a long time if the network is highly congested. If you
    * want to be sure to receive the status of the transaction within a given period you should set
@@ -179,7 +157,7 @@ public class TransactionBuilder {
    * setTimeout</code> does internally; if there's <code>minTime</code> set but no <code>maxTime
    * </code> it will be added). Call to <code>Builder.setTimeout</code> is required if Transaction
    * does not have <code>max_time</code> set. If you don't want to set timeout, use <code>
-   * TIMEOUT_INFINITE</code>. In general you should set <code>TIMEOUT_INFINITE</code> only in smart
+   * TIMEOUT_INFINITE</code>. In general, you should set <code>TIMEOUT_INFINITE</code> only in smart
    * contracts. Please note that Horizon may still return <code>504 Gateway Timeout</code> error,
    * even for short timeouts. In such case you need to resubmit the same transaction again without
    * making any changes to receive a status. This method is using the machine system time (UTC),
@@ -188,35 +166,13 @@ public class TransactionBuilder {
    * @param timeout Timeout in seconds.
    * @return updated Builder
    * @see TimeBounds
-   * @deprecated this method will be removed in upcoming releases, use <code>addPreconditions()
-   *     </code> with TimeBound set instead for more control over preconditions.
    */
-  // TODO: timebounds is the most commonly used precondition, I think
-  //  we can keep this function and not mark it as deprecated?
   public TransactionBuilder setTimeout(BigInteger timeout) {
-    if (preconditions.getTimeBounds() != null
-        && !TIMEOUT_INFINITE.equals(preconditions.getTimeBounds().getMaxTime())) {
-      throw new RuntimeException(
-          "TimeBounds.max_time has been already set - setting timeout would overwrite it.");
-    }
-
     if (timeout.compareTo(BigInteger.ZERO) < 0) {
       throw new RuntimeException("timeout cannot be negative");
     }
 
-    BigInteger timeoutTimestamp =
-        !TIMEOUT_INFINITE.equals(timeout)
-            ? timeout.add(BigInteger.valueOf(System.currentTimeMillis() / 1000L))
-            : TIMEOUT_INFINITE;
-
-    TransactionPreconditionsBuilder preconditionsBuilder = preconditions.toBuilder();
-    if (preconditions.getTimeBounds() == null) {
-      preconditionsBuilder.timeBounds(new TimeBounds(BigInteger.ZERO, timeoutTimestamp));
-    } else {
-      preconditionsBuilder.timeBounds(
-          new TimeBounds(preconditions.getTimeBounds().getMinTime(), timeoutTimestamp));
-    }
-    preconditions = preconditionsBuilder.build();
+    txTimeout = timeout;
     return this;
   }
 
@@ -225,8 +181,6 @@ public class TransactionBuilder {
    *
    * @param timeout Timeout in seconds.
    * @return updated Builder
-   * @deprecated this method will be removed in upcoming releases, use <code>addPreconditions()
-   *     </code> with TimeBound set instead for more control over preconditions.
    */
   public TransactionBuilder setTimeout(long timeout) {
     return setTimeout(BigInteger.valueOf(timeout));
@@ -250,10 +204,25 @@ public class TransactionBuilder {
    * is constructed.
    */
   public Transaction build() {
+    // ensure that the preconditions are valid
+    if (preconditions.getTimeBounds() != null && txTimeout != null) {
+      throw new IllegalArgumentException(
+          "Can not set both TransactionPreconditions.timeBounds and timeout.");
+    }
+
+    if (txTimeout != null) {
+      BigInteger maxTime =
+          !TIMEOUT_INFINITE.equals(txTimeout)
+              ? txTimeout.add(BigInteger.valueOf(System.currentTimeMillis() / 1000L))
+              : TIMEOUT_INFINITE;
+      preconditions =
+          preconditions.toBuilder().timeBounds(new TimeBounds(BigInteger.ZERO, maxTime)).build();
+    }
+
     preconditions.isValid();
 
     if (baseFee == null) {
-      throw new FormatException("mBaseFee has to be set. you must call setBaseFee().");
+      throw new FormatException("baseFee has to be set. you must call setBaseFee().");
     }
 
     if (network == null) {
@@ -284,14 +253,6 @@ public class TransactionBuilder {
    */
   public static Function<TransactionBuilderAccount, Long> IncrementedSequenceNumberFunc =
       TransactionBuilderAccount::getIncrementedSequenceNumber;
-
-  @Deprecated
-  public static org.stellar.sdk.xdr.TimeBounds buildTimeBounds(long minTime, long maxTime) {
-    return new org.stellar.sdk.xdr.TimeBounds.Builder()
-        .minTime(new TimePoint(new Uint64(new XdrUnsignedHyperInteger(minTime))))
-        .maxTime(new TimePoint(new Uint64(new XdrUnsignedHyperInteger(maxTime))))
-        .build();
-  }
 
   /**
    * Sets the transaction's internal Soroban transaction data (resources, footprint, etc.).