+++ src/test/java/org/stellar/sdk/responses/OperationDeserializerTest.java
@@ -11,7 +11,6 @@ import org.junit.Test;
 import org.stellar.sdk.AssetAmount;
 import org.stellar.sdk.AssetTypeNative;
 import org.stellar.sdk.AssetTypePoolShare;
-import org.stellar.sdk.FormatException;
 import org.stellar.sdk.Predicate;
 import org.stellar.sdk.Price;
 import org.stellar.sdk.responses.operations.AccountMergeOperationResponse;
@@ -690,8 +689,7 @@ public class OperationDeserializerTest extends TestCase {
             GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
 
     assertEquals(
-        operation.getSigner().getAccountId(),
-        "GD3ZYXVC7C3ECD5I4E5NGPBFJJSULJ6HJI2FBHGKYFV34DSIWB4YEKJZ");
+        operation.getSignerKey(), "GD3ZYXVC7C3ECD5I4E5NGPBFJJSULJ6HJI2FBHGKYFV34DSIWB4YEKJZ");
     assertEquals(
         operation.getSignerKey(), "GD3ZYXVC7C3ECD5I4E5NGPBFJJSULJ6HJI2FBHGKYFV34DSIWB4YEKJZ");
     assertEquals(operation.getSignerWeight(), new Integer(1));
@@ -705,52 +703,6 @@ public class OperationDeserializerTest extends TestCase {
     assertEquals(operation.getMasterKeyWeight(), new Integer(4));
     assertEquals(operation.getSetFlags()[0], "auth_required_flag");
     assertEquals(operation.getClearFlags()[0], "auth_revocable_flag");
-  }
-
-  @Test
-  public void testDeserializeSetOptionsOperationWithNonEd25519Key() {
-    String json =
-        "{\n"
-            + "        \"_links\": {\n"
-            + "          \"self\": {\n"
-            + "            \"href\": \"https://horizon-testnet.stellar.org/operations/44921793093312513\"\n"
-            + "          },\n"
-            + "          \"transaction\": {\n"
-            + "            \"href\": \"https://horizon-testnet.stellar.org/transactions/d991075183f7740e1aa43700b824f2f404082632f1db9d8a54db00574f83393b\"\n"
-            + "          },\n"
-            + "          \"effects\": {\n"
-            + "            \"href\": \"https://horizon-testnet.stellar.org/operations/44921793093312513/effects\"\n"
-            + "          },\n"
-            + "          \"succeeds\": {\n"
-            + "            \"href\": \"https://horizon-testnet.stellar.org/effects?order=desc\\u0026cursor=44921793093312513\"\n"
-            + "          },\n"
-            + "          \"precedes\": {\n"
-            + "            \"href\": \"https://horizon-testnet.stellar.org/effects?order=asc\\u0026cursor=44921793093312513\"\n"
-            + "          }\n"
-            + "        },\n"
-            + "        \"id\": \"44921793093312513\",\n"
-            + "        \"paging_token\": \"44921793093312513\",\n"
-            + "        \"source_account\": \"GCWYUHCMWC2AATGAXXYZX7T45QZLTRCYNJDD3PC73NEMUXBOCO5F6T6Z\",\n"
-            + "        \"type\": \"set_options\",\n"
-            + "        \"type_i\": 5,\n"
-            + "        \"created_at\": \"2018-08-09T15:36:24Z\",\n"
-            + "        \"transaction_hash\": \"d991075183f7740e1aa43700b824f2f404082632f1db9d8a54db00574f83393b\",\n"
-            + "        \"signer_key\": \"TBGFYVCU76LJ7GZOCGR4X7DG2NV42JPG5CKRL42LA5FZOFI3U2WU7ZAL\",\n"
-            + "        \"signer_weight\": 1\n"
-            + "      }";
-
-    SetOptionsOperationResponse operation =
-        (SetOptionsOperationResponse)
-            GsonSingleton.getInstance().fromJson(json, OperationResponse.class);
-
-    try {
-      operation.getSigner();
-      fail();
-    } catch (FormatException e) {
-      assertEquals("Version byte is invalid", e.getMessage());
-    }
-    assertEquals(
-        operation.getSignerKey(), "TBGFYVCU76LJ7GZOCGR4X7DG2NV42JPG5CKRL42LA5FZOFI3U2WU7ZAL");
   }
 
   @Test